updated regression tests to be query-based. old regression tests have been moved to a separate file to be removed
